+++ src/Http/Controllers/Twoc2pController.php
@@ -11,12 +11,33 @@ class Twoc2pController extends Controller
 {
     public function backend(Request $request)
     {
-        // info('Twoc2p Backend receive.', $request->all());
-        event(new BackendReceived($request->all()));
+        // logger()->info('2C2P Backend : Received', $request->all());
 
-        Twoc2pPayment::create([
-            'action' => Str::after(__METHOD__, '::'),
-            'response' => $request->all()
-        ]);
+        if ($request->payload) {
+            try {
+                $decoded = app('Twoc2p')->decodeJWT($request->payload);
+
+                if ($decoded && data_get($decoded, 'invoiceNo')) {
+                    if (is_object($decoded)) { // if received as object, convert into array
+                        $decoded = json_decode(json_encode($decoded), true);
+                    }
+
+                    event(new BackendReceived($decoded));
+
+                    Twoc2pPayment::create([
+                        'action' => Str::after(__METHOD__, '::'),
+                        'response' => $decoded
+                    ]);
+                } else {
+                    logger()->error('2C2P Backend : Failed to decode token ' . $request->payload);
+                }
+            } catch (\Throwable $th) {
+                //throw $th;
+                logger()->error('2C2P Backend :' . $th->getMessage());
+            }
+        } else {
+            // no payload received
+            logger()->error('2C2P Backend : No payload');
+        }
     }
 }