+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `laraditz/2c2p` will be documented in this file
 
+## 1.0.1 - 2022-07-15
+
+### Added
+- Add `paymentInquiry` method.
+- Add `encodeJWT` method.
+- Add `decodeJWT` method.
+
 ## 1.0.0 - 2021-10-15
 
 - initial release